Updated for shadowgraph

+++ segmentation.py
@@ -1,15 +1,15 @@
-# %%
 import sys
-import argparse
+#import argparse
 import configparser
 from PIL import Image
-import cv2  # still used to save images out
+import cv2
 import os
 import numpy as np
 import csv
 from multiprocessing import Process, Queue
 import tqdm
 import shutil
+import random
 
 class Frame:
     def __init__(self, fpath, name, frame, n):
@@ -59,9 +59,6 @@ def process_frame(q, config): ## TODO: write metadata file
 
     while True:
         frame = q.get()
-        if config['general']['dry_run'] == 'True':
-            print('.')
-            return
         
         ## Read img and flatfield
         gray = cv2.cvtColor(frame.read(), cv2.COLOR_BGR2GRAY)
@@ -133,7 +130,7 @@ def process_avi(avi_path, segmentation_dir, config, q):
             break
 
 
-def process_image(img_path, segmentation_dir, config, q):
+def process_image_dir(img_path, segmentation_dir, config, q):
     """
     This function will take an image folder as input and perform the following steps:
     1. Create output file structures/directories
@@ -144,46 +141,84 @@ def process_image(img_path, segmentation_dir, config, q):
     _, filename = os.path.split(img_path)
     output_path = segmentation_dir + os.path.sep + filename + os.path.sep
     os.makedirs(output_path, exist_ok=True)
-    
 
-    for f in os.listdir(img_path):
-      if f.endswith(".jpg"):
-          video = cv2.VideoCapture(f)
-          if not video.isOpened():
-              return
-        
-          with open(f'{output_path[:-1]} statistics.csv', 'a', newline='\n') as outcsv:
+    with open(f'{output_path[:-1]} statistics.csv', 'a', newline='\n') as outcsv:
               outwritter = csv.writer(outcsv, delimiter=',', quotechar='|')
               outwritter.writerow(['frame', 'crop', 'x', 'y', 'w', 'h'])
+
+    bkg = np.array(cv2.imread(config['segmentation']['calibration_image']))
+        
+    m = np.amin(bkg)
+    M = np.amax(bkg)
+    for f in os.listdir(img_path):
+      if f.endswith(('.jpg', '.jpeg', '.png')):
+          image = np.array(cv2.imread(img_path + os.path.sep + f))
+          image = image / bkg * 255
+          image = image.clip(0,255).astype(np.uint8)
+          #cv2.imwrite(output_path + os.path.sep + f'flat-{n}-{i}.png', image)
+          
+          ## Apply Mask (also invert so black masking becomes white background in shadowgraph)
+          image = ~image
+          mask = np.zeros(image.shape[:2], dtype="uint8")
+          cv2.circle(mask, (image.shape[1]//2, image.shape[0]//2), 1000, 255, -1)
+          image = cv2.bitwise_and(image, image, mask = mask) # Mask
+          #cv2.imwrite(output_path + os.path.sep + f'flatcrop-{n}-{i}.png', ~image)
+          
+          image = ~image # Invert back to shadowgraph-standard
+          q.put(Frame(f, output_path, image, f), block = True)
+
+
+def generate_median_image(directory, output_dir):
+    """
     
-          n = 1
-          while True:
-              ret, frame = video.read()
-              if ret:
-                  q.put(Frame(avi_path, output_path, frame, n), block = True)
-                  n += 1
-              else:
-                  break
+    """
+    # Get a list of all image file names in the directory
+    image_files = [file for file in os.listdir(directory) if file.endswith(('.jpg', '.jpeg', '.png'))]
+    
+    if not image_files:
+        print("No image files found in the directory.")
+        return
 
+    image_files = random.sample(image_files, min([100, len(image_files)]))
+    
+    # Read the first image to get the dimensions
+    first_image_path = os.path.join(directory, image_files[0])
+    first_image = cv2.imread(first_image_path, cv2.IMREAD_GRAYSCALE)
+    height, width = first_image.shape
 
+    # Initialize an array to store all images
+    all_images = np.zeros((len(image_files), height, width), dtype=np.uint8)
+
+    # Load all images into the array
+    for idx, image_file in enumerate(image_files):
+        image_path = os.path.join(directory, image_file)
+        image = cv2.imread(image_path, cv2.IMREAD_GRAYSCALE)
+        image = cv2.GaussianBlur(image, (5,5), 0)
+        all_images[idx] = image
+
+    # Compute the median image
+    median_image = np.median(all_images, axis=0).astype(np.uint8)
+    cv2.imwrite(output_dir + os.path.sep + 'median_image.jpg', median_image)
+    print("New median (calibration) image saved as 'median_image.jpg'.")
+    
 
 if __name__ == "__main__":
     
-    directory = '../../raw/camera0/test1'
+    directory = '../../raw/camera0/shadowgraph'
 
     config = {
         'general' : {
-            'dir_permissions' : 511,
-            'dry_run' : 'False'
+            'dir_permissions' : 511
         },
         'segmentation' : {
             'basename' : 'REG',
             'min_perimeter' : 4*30,
-            'flatfield_q' : 0.05
+            'flatfield_q' : 0.05,
+            'calibration_image' : '../../raw/camera0/shadowgraph/median_image.jpg'
         }
     }
 
-    v_string = "V2024.04.26"
+    v_string = "V2024.05.18"
     print(f"Starting Segmentation Script {v_string}")
 
     ## Determine directories
@@ -197,31 +232,46 @@ if __name__ == "__main__":
     segmentation_dir = segmentation_dir + f"-{config['segmentation']['basename']}" # /media/plankline/Data/analysis/segmentation/Camera1/segmentation/Transect1-REG
     os.makedirs(segmentation_dir, int(config['general']['dir_permissions']), exist_ok = True)
 
-
+    ## Find files to process:
+    # AVI videos
     avis = []
     avis = [os.path.join(raw_dir, avi) for avi in os.listdir(raw_dir) if avi.endswith(".avi")]
     print(f"Number of AVIs found: {len(avis)}")
 
+    # Subfolders of images(?)
+    imgsets = []
+    imgsets = [os.path.join(raw_dir, sub) for sub in next(os.walk(raw_dir))[1]]
+    print(f"Number of possible image sets found: {len(imgsets)}")
+
     ## Prepare workers for receiving frames
     num_threads = os.cpu_count() - 1
-    max_queue = num_threads * 4
+    max_queue = num_threads * 4 # Prepare 4 frames per thread. TODO: test memory vs performance considerations.
     q = Queue(maxsize=int(max_queue))
 
     for i in range(num_threads):
         worker = Process(target=process_frame, args=(q, config,), daemon=True)
         worker.start()
         
-    print(num_threads)
+    print(f'Spun up {num_threads} worker threads for processing.')
 
-    for av in tqdm.tqdm(avis):
-        process_avi(av, segmentation_dir, config, q)
+    if (len(avis)) > 0:
+        print(f'Starting processing on {len(avis)} AVI files.')
+        for av in tqdm.tqdm(avis):
+            process_avi(av, segmentation_dir, config, q)
 
-    for av in avis:
-        _, filename = os.path.split(av)
-        output_path = segmentation_dir + os.path.sep + filename + os.path.sep
-        shutil.make_archive(segmentation_dir + os.path.sep + filename, 'zip', output_path)
-        shutil.rmtree(output_path, ignore_errors=True)
+        for av in avis:
+            _, filename = os.path.split(av)
+            output_path = segmentation_dir + os.path.sep + filename + os.path.sep
+            shutil.make_archive(segmentation_dir + os.path.sep + filename, 'zip', output_path)
+            shutil.rmtree(output_path, ignore_errors=True)
 
+    if len(imgsets) > 0:
+        print(f'Starting processing on {len(imgsets)} subfolders.')
+        for im in tqdm.tqdm(imgsets):
+            if not os.path.exists(config['segmentation']['calibration_image']):
+                generate_median_image(im, config['segmentation']['calibration_image'])
+            process_image_dir(im, segmentation_dir, config, q)
+            
     print('Joining')
     worker.join(timeout=10)
